PDF Reader Re-structure
